Performance just under 2x better for this function.

`demo_nigeria.py` - 6 years with "good" calibration parameters (below)

|System|Before|After|Improvement|
|-------|:-----:|:---:|:-:|
|MacBook Pro M4|86s|32s|2.7x|
|Lenovo X31 Intel iCore 7|412s|144s|2.9x|

```python
best_params = {"r0": 10.645693173058095, "radiation_k": 0.3414061580723608, "seasonal_factor": 0.04574156616477009, "seasonal_phase": 239}

r0 = best_params["r0"]
radiation_k = best_params["radiation_k"]
seasonal_factor = best_params["seasonal_factor"]
seasonal_phase = best_params["seasonal_phase"]
...
    save_pop=True,
    seasonal_factor=seasonal_factor,
    seasonal_phase=seasonal_phase,
)
```