@clorton 2025-05-10
@skroiss-idm - the diff view for this PR is much more confusing than the actual change. The gist is that most of the `Transmission_ABM.step()` function has been consolidated into two Numba functions - one to calculate all the inputs into the infection function, and the infection function (mostly unchanged from `fast_infect()`). The step "prep" function consolidates 3 loops over the population into 1. The updated infect function consolidates a loop over the population _for every node with cases_ into a single loop over the population regardless of the number of nodes with cases. This should be even more impactful for scenarios with ongoing and widespread prevalence compared to the current code.

Since the interface to the prep and infect functions changed, I did not attempt to patch up the alternate infect functions and have removed them from the code since they are now just cluttering `model.py`.